+++ firebase-dataconnect/src/main/kotlin/com/google/firebase/dataconnect/core/DataConnectGrpcClient.kt
@@ -20,6 +20,8 @@ import com.google.firebase.dataconnect.*
 import com.google.firebase.dataconnect.DataConnectPathSegment
 import com.google.firebase.dataconnect.DataSource
 import com.google.firebase.dataconnect.QueryRef.FetchPolicy
+import com.google.firebase.dataconnect.core.DataConnectAppCheck.GetAppCheckTokenResult
+import com.google.firebase.dataconnect.core.DataConnectAuth.GetAuthTokenResult
 import com.google.firebase.dataconnect.core.DataConnectGrpcClientGlobals.toErrorInfoImpl
 import com.google.firebase.dataconnect.core.LoggerGlobals.warn
 import com.google.firebase.dataconnect.util.ProtoUtil.decodeFromStruct
@@ -73,8 +75,9 @@ internal class DataConnectGrpcClient(
     }
 
     val executeQueryResult =
-      grpcRPCs.retryOnGrpcUnauthenticatedError(requestId, "executeQuery") {
-        executeQuery(requestId, request, callerSdkType, fetchPolicy)
+      grpcRPCs.retryOnGrpcUnauthenticatedError(requestId, "executeQuery") { authToken, appCheckToken
+        ->
+        executeQuery(requestId, request, callerSdkType, fetchPolicy, authToken, appCheckToken)
       }
 
     return executeQueryResult.toOperationResult()
@@ -94,7 +97,9 @@ internal class DataConnectGrpcClient(
 
     val response =
       grpcRPCs.retryOnGrpcUnauthenticatedError(requestId, "executeMutation") {
-        executeMutation(requestId, request, callerSdkType)
+        authToken,
+        appCheckToken ->
+        executeMutation(requestId, request, callerSdkType, authToken, appCheckToken)
       }
 
     return OperationResult(
@@ -104,13 +109,16 @@ internal class DataConnectGrpcClient(
     )
   }
 
-  private inline fun <T, R> T.retryOnGrpcUnauthenticatedError(
+  private suspend inline fun <T, R> T.retryOnGrpcUnauthenticatedError(
     requestId: String,
     kotlinMethodName: String,
-    block: T.() -> R
+    block: T.(GetAuthTokenResult?, GetAppCheckTokenResult?) -> R,
   ): R {
+    val authToken1 = dataConnectAuth.getToken(requestId)
+    val appCheckToken1 = dataConnectAppCheck.getToken(requestId)
+
     return try {
-      block()
+      block(authToken1, appCheckToken1)
     } catch (e: StatusException) {
       if (e.status.code != Status.UNAUTHENTICATED.code) {
         throw e
@@ -125,7 +133,10 @@ internal class DataConnectGrpcClient(
       dataConnectAuth.forceRefresh()
       dataConnectAppCheck.forceRefresh()
 
-      block()
+      val authToken2 = dataConnectAuth.getToken(requestId)
+      val appCheckToken2 = dataConnectAppCheck.getToken(requestId)
+
+      block(authToken2, appCheckToken2)
     }
   }
 }