+++ firebase-dataconnect/src/test/kotlin/com/google/firebase/dataconnect/core/DataConnectGrpcClientUnitTest.kt
@@ -24,17 +24,24 @@ import com.google.firebase.dataconnect.DataConnectUntypedData
 import com.google.firebase.dataconnect.DataSource
 import com.google.firebase.dataconnect.FirebaseDataConnect.CallerSdkType
 import com.google.firebase.dataconnect.QueryRef.FetchPolicy
+import com.google.firebase.dataconnect.core.DataConnectAppCheck.GetAppCheckTokenResult
+import com.google.firebase.dataconnect.core.DataConnectAuth.GetAuthTokenResult
 import com.google.firebase.dataconnect.core.DataConnectGrpcClient.OperationResult
 import com.google.firebase.dataconnect.core.DataConnectGrpcClientGlobals.deserialize
 import com.google.firebase.dataconnect.core.DataConnectOperationFailureResponseImpl.ErrorInfoImpl
 import com.google.firebase.dataconnect.testutil.DataConnectLogLevelRule
 import com.google.firebase.dataconnect.testutil.RandomSeedTestRule
 import com.google.firebase.dataconnect.testutil.newMockLogger
+import com.google.firebase.dataconnect.testutil.property.arbitrary.TwoValues
+import com.google.firebase.dataconnect.testutil.property.arbitrary.appCheckTokenResult
+import com.google.firebase.dataconnect.testutil.property.arbitrary.authTokenResult
 import com.google.firebase.dataconnect.testutil.property.arbitrary.dataConnect
 import com.google.firebase.dataconnect.testutil.property.arbitrary.iterator
 import com.google.firebase.dataconnect.testutil.property.arbitrary.operationErrors
+import com.google.firebase.dataconnect.testutil.property.arbitrary.pair
 import com.google.firebase.dataconnect.testutil.property.arbitrary.proto
 import com.google.firebase.dataconnect.testutil.property.arbitrary.struct
+import com.google.firebase.dataconnect.testutil.registerDataConnectKotestPrinters
 import com.google.firebase.dataconnect.testutil.shouldHaveLoggedExactlyOneMessageContaining
 import com.google.firebase.dataconnect.testutil.shouldSatisfy
 import com.google.firebase.dataconnect.util.ProtoUtil.buildStructProto
@@ -74,6 +81,7 @@ import io.kotest.property.arbitrary.int
 import io.kotest.property.arbitrary.map
 import io.kotest.property.arbitrary.merge
 import io.kotest.property.arbitrary.next
+import io.kotest.property.arbitrary.orNull
 import io.kotest.property.arbitrary.string
 import io.kotest.property.assume
 import io.kotest.property.checkAll
@@ -84,7 +92,8 @@ import io.mockk.mockk
 import io.mockk.slot
 import io.mockk.spyk
 import io.mockk.verify
-import java.util.concurrent.atomic.AtomicBoolean
+import java.util.concurrent.atomic.AtomicReference
+import kotlin.random.Random
 import kotlin.reflect.KClass
 import kotlinx.coroutines.test.runTest
 import kotlinx.serialization.DeserializationStrategy
@@ -93,6 +102,7 @@ import kotlinx.serialization.SerializationException
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.modules.SerializersModule
 import kotlinx.serialization.serializer
+import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
@@ -114,15 +124,19 @@ class DataConnectGrpcClientUnitTest {
   private val fetchPolicy = Arb.enum<FetchPolicy>().next(rs)
 
   private val mockDataConnectAuth: DataConnectAuth =
-    mockk(relaxed = true, name = "mockDataConnectAuth-zfbhma6tyh")
+    mockk(relaxed = true, name = "mockDataConnectAuth-zfbhma6tyh") {
+      coEvery { getToken(any()) } returns null
+    }
   private val mockDataConnectAppCheck: DataConnectAppCheck =
-    mockk(relaxed = true, name = "mockDataConnectAppCheck-zfbhma6tyh")
+    mockk(relaxed = true, name = "mockDataConnectAppCheck-zfbhma6tyh") {
+      coEvery { getToken(any()) } returns null
+    }
 
   private val mockDataConnectGrpcRPCs: DataConnectGrpcRPCs =
     mockk(relaxed = true, name = "mockDataConnectGrpcRPCs-zfbhma6tyh") {
-      coEvery { executeQuery(any(), any(), any(), any()) } returns
+      coEvery { executeQuery(any(), any(), any(), any(), any(), any()) } returns
         DataConnectGrpcRPCs.ExecuteQueryResult.FromServer(ExecuteQueryResponse.getDefaultInstance())
-      coEvery { executeMutation(any(), any(), any()) } returns
+      coEvery { executeMutation(any(), any(), any(), any(), any()) } returns
         ExecuteMutationResponse.getDefaultInstance()
     }
 
@@ -137,6 +151,11 @@ class DataConnectGrpcClientUnitTest {
       dataConnectAppCheck = mockDataConnectAppCheck,
       logger = mockLogger,
     )
+
+  @Before
+  fun registerPrinters() {
+    registerDataConnectKotestPrinters()
+  }
 
   @Test
   fun `executeQuery() should forward requestId, callerSdkType, and fetchPolicy`() = runTest {
@@ -154,7 +173,14 @@ class DataConnectGrpcClientUnitTest {
         fetchPolicy
       )
       coVerify {
-        mockDataConnectGrpcRPCs.executeQuery(requestId, any(), callerSdkType, fetchPolicy)
+        mockDataConnectGrpcRPCs.executeQuery(
+          requestId,
+          any(),
+          callerSdkType,
+          fetchPolicy,
+          any(),
+          any()
+        )
       }
     }
   }
@@ -181,7 +207,9 @@ class DataConnectGrpcClientUnitTest {
           .setOperationName(operationName)
           .setVariables(variables)
           .build()
-      coVerify { mockDataConnectGrpcRPCs.executeQuery(any(), expectedRequest, any(), any()) }
+      coVerify {
+        mockDataConnectGrpcRPCs.executeQuery(any(), expectedRequest, any(), any(), any(), any())
+      }
     }
   }
 
@@ -192,7 +220,9 @@ class DataConnectGrpcClientUnitTest {
       callerSdkType ->
       dataConnectGrpcClient.executeMutation(requestId, operationName, variables, callerSdkType)
 
-      coVerify { mockDataConnectGrpcRPCs.executeMutation(requestId, any(), callerSdkType) }
+      coVerify {
+        mockDataConnectGrpcRPCs.executeMutation(requestId, any(), callerSdkType, any(), any())
+      }
     }
   }
 
@@ -211,14 +241,17 @@ class DataConnectGrpcClientUnitTest {
         .setOperationName(operationName)
         .setVariables(variables)
         .build()
-    coVerify { mockDataConnectGrpcRPCs.executeMutation(any(), expectedRequest, any()) }
+    coVerify {
+      mockDataConnectGrpcRPCs.executeMutation(any(), expectedRequest, any(), any(), any())
+    }
   }
 
   @Test
   fun `executeQuery() should return data and empty errors if response is from cache`() = runTest {
     val responseData = Arb.proto.struct().next(rs).struct
-    coEvery { mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any()) } returns
-      DataConnectGrpcRPCs.ExecuteQueryResult.FromCache(responseData)
+    coEvery {
+      mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any(), any(), any())
+    } returns DataConnectGrpcRPCs.ExecuteQueryResult.FromCache(responseData)
 
     val operationResult =
       dataConnectGrpcClient.executeQuery(
@@ -235,7 +268,9 @@ class DataConnectGrpcClientUnitTest {
 
   @Test
   fun `executeQuery() should return null data and empty errors if response is empty`() = runTest {
-    coEvery { mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any()) } returns
+    coEvery {
+      mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any(), any(), any())
+    } returns
       DataConnectGrpcRPCs.ExecuteQueryResult.FromServer(ExecuteQueryResponse.getDefaultInstance())
 
     val operationResult =
@@ -253,7 +288,7 @@ class DataConnectGrpcClientUnitTest {
   @Test
   fun `executeMutation() should return null data and empty errors if response is empty`() =
     runTest {
-      coEvery { mockDataConnectGrpcRPCs.executeMutation(any(), any(), any()) } returns
+      coEvery { mockDataConnectGrpcRPCs.executeMutation(any(), any(), any(), any(), any()) } returns
         ExecuteMutationResponse.getDefaultInstance()
 
       val operationResult =
@@ -266,7 +301,9 @@ class DataConnectGrpcClientUnitTest {
   fun `executeQuery() should return data and errors`() = runTest {
     val responseData = Arb.proto.struct().next(rs).struct
     val responseErrors = List(3) { GraphqlErrorInfo.random(RandomSource.default()) }
-    coEvery { mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any()) } returns
+    coEvery {
+      mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any(), any(), any())
+    } returns
       DataConnectGrpcRPCs.ExecuteQueryResult.FromServer(
         ExecuteQueryResponse.newBuilder()
           .setData(responseData)
@@ -295,7 +332,7 @@ class DataConnectGrpcClientUnitTest {
   fun `executeMutation() should return data and errors`() = runTest {
     val responseData = Arb.proto.struct().next(rs).struct
     val responseErrors = List(3) { GraphqlErrorInfo.random(RandomSource.default()) }
-    coEvery { mockDataConnectGrpcRPCs.executeMutation(any(), any(), any()) } returns
+    coEvery { mockDataConnectGrpcRPCs.executeMutation(any(), any(), any(), any(), any()) } returns
       ExecuteMutationResponse.newBuilder()
         .setData(responseData)
         .addAllErrors(responseErrors.map { it.graphqlError })
@@ -315,7 +352,9 @@ class DataConnectGrpcClientUnitTest {
   @Test
   fun `executeQuery() should propagate non-grpc exceptions`() = runTest {
     val exception = TestException("k6hzgp7hvz")
-    coEvery { mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any()) } throws exception
+    coEvery {
+      mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any(), any(), any())
+    } throws exception
 
     val thrownException =
       shouldThrow<TestException> {
@@ -334,7 +373,8 @@ class DataConnectGrpcClientUnitTest {
   @Test
   fun `executeMutation() should propagate non-grpc exceptions`() = runTest {
     val exception = TestException("g32376rnd3")
-    coEvery { mockDataConnectGrpcRPCs.executeMutation(any(), any(), any()) } throws exception
+    coEvery { mockDataConnectGrpcRPCs.executeMutation(any(), any(), any(), any(), any()) } throws
+      exception
 
     val thrownException =
       shouldThrow<TestException> {
@@ -345,149 +385,108 @@ class DataConnectGrpcClientUnitTest {
   }
 
   @Test
-  fun `executeQuery() should retry with a fresh auth token on UNAUTHENTICATED`() = runTest {
-    val responseData = Arb.proto.struct().next(rs).struct
-    val forceRefresh = AtomicBoolean(false)
-    coEvery { mockDataConnectAuth.forceRefresh() } answers { forceRefresh.set(true) }
-    coEvery { mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any()) } answers
-      {
-        if (forceRefresh.get()) {
-          DataConnectGrpcRPCs.ExecuteQueryResult.FromServer(
-            ExecuteQueryResponse.newBuilder().setData(responseData).build()
-          )
-        } else {
-          // Use a custom description to ensure that DataConnectGrpcClient is checking for just
-          // the code, and not the entire equality of Status.UNAUTHENTICATED.
-          throw StatusException(
-            Status.UNAUTHENTICATED.withDescription(
-              "this error should be ignored and result in a retry with a fresh token n2ak4cq6jr"
-            )
-          )
-        }
-      }
+  fun `executeQuery() should retry with fresh auth and app check tokens on UNAUTHENTICATED`() =
+    runTest {
+      val authTokens = authTokenPairArb().next(rs)
+      mockDataConnectAuth.stubGetTokensSimulatingForceRefresh(authTokens)
+      val appCheckTokens = appCheckTokenPairArb().next(rs)
+      mockDataConnectAppCheck.stubGetTokensSimulatingForceRefresh(appCheckTokens)
+      coEvery {
+        mockDataConnectGrpcRPCs.executeQuery(
+          any(),
+          any(),
+          any(),
+          any(),
+          matchNullable { it == authTokens.value1 },
+          matchNullable { it == appCheckTokens.value1 },
+        )
+      } throws
+        StatusException(
+          Status.UNAUTHENTICATED.withDescription("status exception ${Random.nextInt()} bjh9zc6h5n")
+        )
+      val responseData = Arb.proto.struct().next(rs).struct
+      coEvery {
+        mockDataConnectGrpcRPCs.executeQuery(
+          any(),
+          any(),
+          any(),
+          any(),
+          matchNullable { it == authTokens.value2 },
+          matchNullable { it == appCheckTokens.value2 },
+        )
+      } returns
+        DataConnectGrpcRPCs.ExecuteQueryResult.FromServer(
+          ExecuteQueryResponse.newBuilder().setData(responseData).build()
+        )
 
-    val result =
-      dataConnectGrpcClient.executeQuery(
-        requestId,
-        operationName,
-        variables,
-        callerSdkType,
-        fetchPolicy
+      val result =
+        dataConnectGrpcClient.executeQuery(
+          requestId,
+          operationName,
+          variables,
+          callerSdkType,
+          fetchPolicy
+        )
+
+      result shouldBe OperationResult(data = responseData, errors = emptyList(), DataSource.SERVER)
+      coVerify(exactly = 2) {
+        mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any(), any(), any())
+      }
+      mockLogger.shouldHaveLoggedExactlyOneMessageContaining(
+        "retrying with fresh Auth and/or AppCheck tokens"
       )
-
-    result shouldBe OperationResult(data = responseData, errors = emptyList(), DataSource.SERVER)
-    coVerify(exactly = 2) { mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any()) }
-    mockLogger.shouldHaveLoggedExactlyOneMessageContaining(
-      "retrying with fresh Auth and/or AppCheck tokens"
-    )
-    mockLogger.shouldHaveLoggedExactlyOneMessageContaining("UNAUTHENTICATED")
-  }
+      mockLogger.shouldHaveLoggedExactlyOneMessageContaining("UNAUTHENTICATED")
+    }
 
   @Test
-  fun `executeMutation() should retry with a fresh auth token on UNAUTHENTICATED`() = runTest {
-    val responseData = Arb.proto.struct().next(rs).struct
-    val forceRefresh = AtomicBoolean(false)
-    coEvery { mockDataConnectAuth.forceRefresh() } answers { forceRefresh.set(true) }
-    coEvery { mockDataConnectGrpcRPCs.executeMutation(any(), any(), any()) } answers
-      {
-        if (forceRefresh.get()) {
-          ExecuteMutationResponse.newBuilder().setData(responseData).build()
-        } else {
-          // Use a custom description to ensure that DataConnectGrpcClient is checking for just
-          // the code, and not the entire equality of Status.UNAUTHENTICATED.
-          throw StatusException(
-            Status.UNAUTHENTICATED.withDescription(
-              "this error should be ignored and result in a retry with a fresh token p3vmc3gs5v"
-            )
-          )
-        }
+  fun `executeMutation() should retry with fresh auth and app check tokens on UNAUTHENTICATED`() =
+    runTest {
+      val authTokens = authTokenPairArb().next(rs)
+      mockDataConnectAuth.stubGetTokensSimulatingForceRefresh(authTokens)
+      val appCheckTokens = appCheckTokenPairArb().next(rs)
+      mockDataConnectAppCheck.stubGetTokensSimulatingForceRefresh(appCheckTokens)
+      coEvery {
+        mockDataConnectGrpcRPCs.executeMutation(
+          any(),
+          any(),
+          any(),
+          matchNullable { it == authTokens.value1 },
+          matchNullable { it == appCheckTokens.value1 },
+        )
+      } throws
+        StatusException(
+          Status.UNAUTHENTICATED.withDescription("status exception ${Random.nextInt()} m8gzej7pmy")
+        )
+      val responseData = Arb.proto.struct().next(rs).struct
+      coEvery {
+        mockDataConnectGrpcRPCs.executeMutation(
+          any(),
+          any(),
+          any(),
+          matchNullable { it == authTokens.value2 },
+          matchNullable { it == appCheckTokens.value2 },
+        )
+      } returns ExecuteMutationResponse.newBuilder().setData(responseData).build()
+
+      val result =
+        dataConnectGrpcClient.executeMutation(requestId, operationName, variables, callerSdkType)
+
+      result shouldBe OperationResult(data = responseData, errors = emptyList(), DataSource.SERVER)
+      coVerify(exactly = 2) {
+        mockDataConnectGrpcRPCs.executeMutation(any(), any(), any(), any(), any())
       }
-
-    val result =
-      dataConnectGrpcClient.executeMutation(requestId, operationName, variables, callerSdkType)
-
-    result shouldBe OperationResult(data = responseData, errors = emptyList(), DataSource.SERVER)
-    coVerify(exactly = 2) { mockDataConnectGrpcRPCs.executeMutation(any(), any(), any()) }
-    mockLogger.shouldHaveLoggedExactlyOneMessageContaining(
-      "retrying with fresh Auth and/or AppCheck tokens"
-    )
-    mockLogger.shouldHaveLoggedExactlyOneMessageContaining("UNAUTHENTICATED")
-  }
-
-  @Test
-  fun `executeQuery() should retry with a fresh AppCheck token on UNAUTHENTICATED`() = runTest {
-    val responseData = Arb.proto.struct().next(rs).struct
-    val forceRefresh = AtomicBoolean(false)
-    coEvery { mockDataConnectAppCheck.forceRefresh() } answers { forceRefresh.set(true) }
-    coEvery { mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any()) } answers
-      {
-        if (forceRefresh.get()) {
-          DataConnectGrpcRPCs.ExecuteQueryResult.FromServer(
-            ExecuteQueryResponse.newBuilder().setData(responseData).build()
-          )
-        } else {
-          // Use a custom description to ensure that DataConnectGrpcClient is checking for just
-          // the code, and not the entire equality of Status.UNAUTHENTICATED.
-          throw StatusException(
-            Status.UNAUTHENTICATED.withDescription(
-              "this error should be ignored and result in a retry with a fresh token tepb5xq4kk"
-            )
-          )
-        }
-      }
-
-    val result =
-      dataConnectGrpcClient.executeQuery(
-        requestId,
-        operationName,
-        variables,
-        callerSdkType,
-        fetchPolicy
+      mockLogger.shouldHaveLoggedExactlyOneMessageContaining(
+        "retrying with fresh Auth and/or AppCheck tokens"
       )
-
-    result shouldBe OperationResult(data = responseData, errors = emptyList(), DataSource.SERVER)
-    coVerify(exactly = 2) { mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any()) }
-    mockLogger.shouldHaveLoggedExactlyOneMessageContaining(
-      "retrying with fresh Auth and/or AppCheck tokens"
-    )
-    mockLogger.shouldHaveLoggedExactlyOneMessageContaining("UNAUTHENTICATED")
-  }
-
-  @Test
-  fun `executeMutation() should retry with a fresh AppCheck token on UNAUTHENTICATED`() = runTest {
-    val responseData = Arb.proto.struct().next(rs).struct
-    val forceRefresh = AtomicBoolean(false)
-    coEvery { mockDataConnectAppCheck.forceRefresh() } answers { forceRefresh.set(true) }
-    coEvery { mockDataConnectGrpcRPCs.executeMutation(any(), any(), any()) } answers
-      {
-        if (forceRefresh.get()) {
-          ExecuteMutationResponse.newBuilder().setData(responseData).build()
-        } else {
-          // Use a custom description to ensure that DataConnectGrpcClient is checking for just
-          // the code, and not the entire equality of Status.UNAUTHENTICATED.
-          throw StatusException(
-            Status.UNAUTHENTICATED.withDescription(
-              "this error should be ignored and result in a retry with a fresh token v2449h6ty8"
-            )
-          )
-        }
-      }
-
-    val result =
-      dataConnectGrpcClient.executeMutation(requestId, operationName, variables, callerSdkType)
-
-    result shouldBe OperationResult(data = responseData, errors = emptyList(), DataSource.SERVER)
-    coVerify(exactly = 2) { mockDataConnectGrpcRPCs.executeMutation(any(), any(), any()) }
-    mockLogger.shouldHaveLoggedExactlyOneMessageContaining(
-      "retrying with fresh Auth and/or AppCheck tokens"
-    )
-    mockLogger.shouldHaveLoggedExactlyOneMessageContaining("UNAUTHENTICATED")
-  }
+      mockLogger.shouldHaveLoggedExactlyOneMessageContaining("UNAUTHENTICATED")
+    }
 
   @Test
   fun `executeQuery() should NOT retry on error status other than UNAUTHENTICATED`() = runTest {
     val exception = StatusException(Status.INTERNAL)
-    coEvery { mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any()) } throws exception
+    coEvery {
+      mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any(), any(), any())
+    } throws exception
 
     val thrownException =
       shouldThrow<StatusException> {
@@ -508,7 +507,8 @@ class DataConnectGrpcClientUnitTest {
   @Test
   fun `executeMutation() should NOT retry on error status other than UNAUTHENTICATED`() = runTest {
     val exception = StatusException(Status.INTERNAL)
-    coEvery { mockDataConnectGrpcRPCs.executeMutation(any(), any(), any()) } throws exception
+    coEvery { mockDataConnectGrpcRPCs.executeMutation(any(), any(), any(), any(), any()) } throws
+      exception
 
     val thrownException =
       shouldThrow<StatusException> {
@@ -525,8 +525,9 @@ class DataConnectGrpcClientUnitTest {
     runTest {
       val exception1 = StatusException(Status.UNAUTHENTICATED)
       val exception2 = StatusException(Status.UNAUTHENTICATED)
-      coEvery { mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any()) } throwsMany
-        (listOf(exception1, exception2))
+      coEvery {
+        mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any(), any(), any())
+      } throwsMany (listOf(exception1, exception2))
 
       val thrownException =
         shouldThrow<StatusException> {
@@ -547,8 +548,9 @@ class DataConnectGrpcClientUnitTest {
     runTest {
       val exception1 = StatusException(Status.UNAUTHENTICATED)
       val exception2 = StatusException(Status.UNAUTHENTICATED)
-      coEvery { mockDataConnectGrpcRPCs.executeMutation(any(), any(), any()) } throwsMany
-        (listOf(exception1, exception2))
+      coEvery {
+        mockDataConnectGrpcRPCs.executeMutation(any(), any(), any(), any(), any())
+      } throwsMany (listOf(exception1, exception2))
 
       val thrownException =
         shouldThrow<StatusException> {
@@ -563,8 +565,9 @@ class DataConnectGrpcClientUnitTest {
     runTest {
       val exception1 = StatusException(Status.UNAUTHENTICATED)
       val exception2 = StatusException(Status.ABORTED)
-      coEvery { mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any()) } throwsMany
-        (listOf(exception1, exception2))
+      coEvery {
+        mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any(), any(), any())
+      } throwsMany (listOf(exception1, exception2))
 
       val thrownException =
         shouldThrow<StatusException> {
@@ -585,8 +588,9 @@ class DataConnectGrpcClientUnitTest {
     runTest {
       val exception1 = StatusException(Status.UNAUTHENTICATED)
       val exception2 = StatusException(Status.ABORTED)
-      coEvery { mockDataConnectGrpcRPCs.executeMutation(any(), any(), any()) } throwsMany
-        (listOf(exception1, exception2))
+      coEvery {
+        mockDataConnectGrpcRPCs.executeMutation(any(), any(), any(), any(), any())
+      } throwsMany (listOf(exception1, exception2))
 
       val thrownException =
         shouldThrow<StatusException> {
@@ -601,8 +605,9 @@ class DataConnectGrpcClientUnitTest {
     runTest {
       val exception1 = StatusException(Status.UNAUTHENTICATED)
       val exception2 = TestException("eysrmxmxk7")
-      coEvery { mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any()) } throwsMany
-        (listOf(exception1, exception2))
+      coEvery {
+        mockDataConnectGrpcRPCs.executeQuery(any(), any(), any(), any(), any(), any())
+      } throwsMany (listOf(exception1, exception2))
 
       val thrownException =
         shouldThrow<TestException> {
@@ -623,8 +628,9 @@ class DataConnectGrpcClientUnitTest {
     runTest {
       val exception1 = StatusException(Status.UNAUTHENTICATED)
       val exception2 = TestException("qz2ykb8wa2")
-      coEvery { mockDataConnectGrpcRPCs.executeMutation(any(), any(), any()) } throwsMany
-        (listOf(exception1, exception2))
+      coEvery {
+        mockDataConnectGrpcRPCs.executeMutation(any(), any(), any(), any(), any())
+      } throwsMany (listOf(exception1, exception2))
 
       val thrownException =
         shouldThrow<TestException> {
@@ -908,3 +914,27 @@ class DataConnectGrpcClientOperationResultUnitTest {
     }
   }
 }
+
+private fun DataConnectAuth.stubGetTokensSimulatingForceRefresh(
+  tokens: Iterable<GetAuthTokenResult?>
+) {
+  val tokenIterator = tokens.iterator()
+  val currentToken = AtomicReference(tokenIterator.next())
+  coEvery { getToken(any()) } answers { currentToken.get() }
+  coEvery { forceRefresh() } answers { currentToken.set(tokenIterator.next()) }
+}
+
+private fun DataConnectAppCheck.stubGetTokensSimulatingForceRefresh(
+  tokens: Iterable<GetAppCheckTokenResult?>
+) {
+  val tokenIterator = tokens.iterator()
+  val currentToken = AtomicReference(tokenIterator.next())
+  coEvery { getToken(any()) } answers { currentToken.get() }
+  coEvery { forceRefresh() } answers { currentToken.set(tokenIterator.next()) }
+}
+
+private fun authTokenPairArb(): Arb<TwoValues<GetAuthTokenResult?>> =
+  Arb.dataConnect.authTokenResult().orNull(nullProbability = 0.33).pair()
+
+private fun appCheckTokenPairArb(): Arb<TwoValues<GetAppCheckTokenResult?>> =
+  Arb.dataConnect.appCheckTokenResult().orNull(nullProbability = 0.33).pair()