+++ firebase-dataconnect/CHANGELOG.md
@@ -11,6 +11,8 @@
   ([#8024](https://github.com/firebase/firebase-android-sdk/pull/8024))
 - [changed] Internal refactor to use more descriptive variable names.
   ([#8025](https://github.com/firebase/firebase-android-sdk/pull/8025))
+- [changed] Internal refactor to use token objects instead of strings.
+  ([#8027](https://github.com/firebase/firebase-android-sdk/pull/8027))
 
 # 17.2.0